+++ tests/Core/OutputTest.lua
@@ -1,14 +1,10 @@
-TestOutput = {}
+TestOutput = BaseTestClass:new()
     -- @covers Output:color()
     function TestOutput:testColor()
         local function execution(value, color, primaryColor, expectedOutput)
-            local originalPrimaryColor = __.addon.colors.primary
-
             __.addon.colors.primary = primaryColor
 
             lu.assertEquals(expectedOutput, __.output:color(value, color))
-
-            __.addon.colors.primary = originalPrimaryColor
         end
 
         execution('test', nil, nil, 'test')