+++ tests/Facades/EventHandlers/PlayerLoginEventHandlerTest.lua
@@ -1,4 +1,4 @@
-TestPlayerLoginEventHandler = {}
+TestPlayerLoginEventHandler = BaseTestClass:new()
     -- @covers PlayerLoginEventHandler.lua
     function TestPlayerLoginEventHandler:testEventNameIsSet()
         lu.assertEquals('PLAYER_LOGIN', __.events.EVENT_NAME_PLAYER_LOGIN)