+++ documentation/docs/changelog.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## 2024.mm.yy - version 0.0.7-alpha
+## 2024.04.03 - version 0.0.8-alpha
+
+* Add a new support class to handle bool values
+* Add new support methods to Str
+* Reset the library instance for each test method for better mocking and less garbage between test cases
+
+## 2024.03.29 - version 0.0.7-alpha
 
 * Add the Events facade class to serve as a way to improve event handling
     * Listen to `PLAYER_LOGIN`